pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Anna Kudriaeva <anna.kudriaeva@gmail.com>"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
jinja2 = "^3.1.4"
validators = "^0.33.0"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"
black = "^24.10.0"
ruff = "^0.8.1"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
